Introduced NUMA-awareness and support for inputs with more than 2^32 rows
NUMA awareness and multiple memory transfer performance fixes.
Fix on input size data type to handle bigger inputs.
Temporary fix (see #2) for a race condition on (device) buffer
management (replaced the lockless stack with a spin-lock-protected
stack). Temporarily switched off materialization (results are still
transfered to the CPU) (opens #3 as a reminder).

+++ src/operators/generators.cuh
@@ -13,9 +13,9 @@ private:
     cudaStream_t                        strm;
     buffer_pool<int32_t>::buffer_t    **buff_ret;
     int32_t                            *src;
-    uint32_t                            N;
+    size_t                              N;
 public:
-    __host__ generator(h_operator_t * parent, int32_t *src, uint32_t N);
+    __host__ generator(h_operator_t * parent, int32_t *src, size_t N);
 
     __host__ void open();
